Substitute Int8 for non-portable "long long"
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@13036 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/blast/api/blast_options.hpp
@@ -576,7 +576,7 @@ private:
         m_ReqOpts->Set().push_back(p);
     }
     
-    void x_SetOneParam(const char * name, const long long int * x)
+    void x_SetOneParam(const char * name, const Int8 * x)
     {
         CRef<objects::CBlast4_value> v(new objects::CBlast4_value);
         v->SetBig_integer(*x);
@@ -2414,6 +2414,9 @@ END_NCBI_SCOPE
 * ===========================================================================
 *
 * $Log$
+* Revision 1.39  2004/01/17 00:15:06  ucko
+* Substitute Int8 for non-portable "long long"
+*
 * Revision 1.38  2004/01/16 21:40:01  bealer
 * - Add Blast4 API support to the CBlastOptions class.
 *
